Let stack take ownership of data
Instead of having the commands keep the data as &mut, *mut or
using interior mutability, the stack now takes ownership, as
proposed in #1.
This makes the API easier to work with since interior mutability
is not needed, and borrowing is now only used on method calls.

Also rename the types to Command, Stack and Group.